requirements.txt

streamlit
google-generativeai
openai-whisper
soundfile
